move step counter to optimizer state and refactor chat generation
  move step counter from global variable to Muon optimizer state field.
  update checkpointing to handle step counter with float32 conversion.
  rewrite chat_cli generation to use padded buffers with JIT-compiled prefill.
  preserve dtypes in model updates across all training scripts.

+++ nanojax/gpt.py
@@ -180,7 +180,7 @@ class GPT:
         # with the benefit of finer-grained control over mixed precision.
         
         # project our tokens into embedding space
-        x = self.wte[idx].astype(compute_dtype)
+        x = self.wte[idx]
         x = rms_norm(x)
 
         h = self.cfg.n_embed // self.cfg.n_head
